WIP: Links
Giving up for today. The thing I'm struggling with right now is where to update
the internal URL of _links and _embedded resources. The self links are at
different places. One approach would be to give link resourcesa a fake _links
attribute with a self reference. With the current implementation, this would
lead to infinite recursions, though. The other way I can think of right now is
to give (Lazy)Resource a hint where to find the URL or just give it to it from
the start.