feat(lib): add `related` method on Resource
The `#related` method allows navigation using embedded or links.

This enables consumers to not care in which section the HAL the link is
placed. It looks in embedded first, and if the rel is not found there
it then looks in links. Templated links are expanded using the
optional `params` argument.